ROX-11169: Remove external plugin e2e test
