+++ qa-tests-backend/src/test/groovy/SACTest.groovy
@@ -1,4 +1,8 @@
 import static Services.waitForViolation
+import static io.stackrox.proto.storage.RoleOuterClass.Access.NO_ACCESS
+import static io.stackrox.proto.storage.RoleOuterClass.Access.READ_ACCESS
+import static io.stackrox.proto.storage.RoleOuterClass.Access.READ_WRITE_ACCESS
+import static io.stackrox.proto.storage.RoleOuterClass.SimpleAccessScope.newBuilder
 import static services.ClusterService.DEFAULT_CLUSTER_NAME
 
 import orchestratormanager.OrchestratorTypes
@@ -7,6 +11,7 @@ import io.stackrox.proto.api.v1.ApiTokenService.GenerateTokenResponse
 import io.stackrox.proto.api.v1.NamespaceServiceOuterClass
 import io.stackrox.proto.api.v1.SearchServiceOuterClass as SSOC
 import io.stackrox.proto.storage.DeploymentOuterClass
+import io.stackrox.proto.storage.RoleOuterClass
 
 import groups.BAT
 import objects.Deployment
@@ -17,6 +22,7 @@ import services.DeploymentService
 import services.ImageService
 import services.NamespaceService
 import services.NetworkGraphService
+import services.RoleService
 import services.SearchService
 import services.SecretService
 import services.SummaryService
@@ -26,6 +32,7 @@ import util.NetworkGraphUtil
 import org.junit.AssumptionViolatedException
 import org.junit.experimental.categories.Category
 import spock.lang.IgnoreIf
+import spock.lang.Shared
 import spock.lang.Unroll
 
 @Category(BAT)
@@ -35,7 +42,6 @@ class SACTest extends BaseSpecification {
     static final private String NAMESPACE_QA1 = "qa-test1"
     static final private String DEPLOYMENTNGINX_NAMESPACE_QA2 = "sac-deploymentnginx-qa2"
     static final private String NAMESPACE_QA2 = "qa-test2"
-    static final private String TESTROLE = "Continuous Integration"
     static final private String SECRETNAME = "sac-secret"
     static final protected String ALLACCESSTOKEN = "allAccessToken"
     static final protected String NOACCESSTOKEN = "noAccess"
@@ -71,6 +77,12 @@ class SACTest extends BaseSpecification {
     static final private Integer WAIT_FOR_RISK_RETRIES =
             isRaceBuild() ? 300 : ((Env.mustGetOrchestratorType() == OrchestratorTypes.OPENSHIFT) ? 80 : 50)
 
+    @Shared
+    private Map<String, RoleOuterClass.Access> allResourcesAccess
+
+    @Shared
+    private Map<String, List<String>> tokenToRoles
+
     def setup() {
         BaseService.useBasicAuth()
     }
@@ -100,6 +112,45 @@ class SACTest extends BaseSpecification {
                 throw new AssumptionViolatedException("Failed to retrieve risk from deployment ${dep.name}", e)
             }
         }
+
+        disableAuthzPlugin()
+
+        allResourcesAccess = RoleService.resources.resourcesList.collectEntries { [it, READ_WRITE_ACCESS] }
+
+        // TODO: Replace with the defaultAccessScope id: "denyall"
+        def noaccessScope = RoleService.createAccessScope(newBuilder()
+                .setName("no-access-scope").build())
+        def remoteQaTest1 = createAccessScope(DEFAULT_CLUSTER_NAME, "qa-test1")
+        def remoteQaTest2 = createAccessScope(DEFAULT_CLUSTER_NAME, "qa-test2")
+
+        def noaccess = createRole(noaccessScope.id, allResourcesAccess)
+
+        tokenToRoles = [
+                (NOACCESSTOKEN)                   : [noaccess],
+                (ALLACCESSTOKEN)                  : [createRole(UNRESTRICTED_SCOPE_ID, allResourcesAccess)],
+                "deployments-access-token"        : [createRole(remoteQaTest2.id,
+                        ["Deployment": READ_ACCESS, "Risk": READ_ACCESS])],
+                "getSummaryCountsToken"           : [createRole(remoteQaTest1.id, allResourcesAccess)],
+                "listSecretsToken"                : [createRole(UNRESTRICTED_SCOPE_ID, ["Secret": READ_ACCESS])],
+                "searchAlertsToken"               : [createRole(remoteQaTest1.id, ["Alert": READ_ACCESS]), noaccess],
+                "searchDeploymentsToken"          : [createRole(remoteQaTest1.id,
+                        ["Deployment": READ_ACCESS]), noaccess],
+                "searchImagesToken"               : [createRole(remoteQaTest1.id, ["Image": READ_ACCESS]), noaccess],
+                "searchNamespacesToken"           : [createRole(remoteQaTest1.id,
+                        ["Namespace": READ_ACCESS]), noaccess],
+                "searchDeploymentsImagesToken"    : [createRole(remoteQaTest1.id,
+                        ["Deployment": READ_ACCESS, "Image": READ_ACCESS]), noaccess],
+                "stackroxNetFlowsToken"           : [createRole(createAccessScope(DEFAULT_CLUSTER_NAME, "stackrox").id,
+                        ["Deployment": READ_ACCESS, "NetworkGraph": READ_ACCESS]),
+                                                     createRole(UNRESTRICTED_SCOPE_ID, ["Cluster": READ_ACCESS]),
+                                                     noaccess],
+                "kubeSystemDeploymentsImagesToken": [createRole(createAccessScope(
+                        DEFAULT_CLUSTER_NAME, "kube-system").id, ["Deployment": READ_ACCESS, "Image": READ_ACCESS]),
+                                                     noaccess],
+                "aggregatedToken"                 : [createRole(remoteQaTest2.id, ["Deployment": READ_ACCESS]),
+                                                     createRole(remoteQaTest1.id, ["Deployment": NO_ACCESS]),
+                                                     noaccess],
+        ]
     }
 
     def cleanupSpec() {
@@ -112,6 +163,7 @@ class SACTest extends BaseSpecification {
                 orchestrator.deleteNamespace(ns)
                 orchestrator.waitForNamespaceDeletion(ns)
         }
+        cleanupRole(*(tokenToRoles.values().flatten().unique()))
     }
 
     static getAlertCount() {
@@ -131,7 +183,7 @@ class SACTest extends BaseSpecification {
     }
 
     GenerateTokenResponse useToken(String tokenName) {
-        GenerateTokenResponse token = ApiTokenService.generateToken(tokenName, TESTROLE)
+        GenerateTokenResponse token = ApiTokenService.generateToken(tokenName, *(tokenToRoles.get(tokenName)))
         BaseService.useApiToken(token.token)
         token
     }
@@ -153,7 +205,33 @@ class SACTest extends BaseSpecification {
         orchestrator.deleteSecret(SECRETNAME, namespace)
     }
 
-    Boolean summaryTestShouldSeeNoClustersAndNodes() { true }
+    def cleanupRole(String... roleName) {
+        roleName.each {
+            try {
+                def role = RoleService.getRole(it)
+                RoleService.deleteRole(role.name)
+                RoleService.deleteAccessScope(role.accessScopeId)
+            } catch (Exception e) {
+                log.error("Error deleting role ${name}", e)
+            }
+        }
+    }
+
+    String createRole(String sacId, Map<String, RoleOuterClass.Access> resources) {
+        RoleService.createRoleWithScopeAndPermissionSet("SACv2 Test Automation Role " + UUID.randomUUID(),
+                sacId, resources
+        ).name
+    }
+
+    def createAccessScope(String clusterName, String namespaceName) {
+        RoleService.createAccessScope(newBuilder()
+                .setName(UUID.randomUUID().toString())
+                .setRules(RoleOuterClass.SimpleAccessScope.Rules.newBuilder()
+                        .addIncludedNamespaces(RoleOuterClass.SimpleAccessScope.Rules.Namespace.newBuilder()
+                                .setClusterName(clusterName)
+                                .setNamespaceName(namespaceName)))
+                .build())
+    }
 
     @Unroll
     @IgnoreIf({ Env.CI_JOBNAME.contains("postgres") })
@@ -208,10 +286,6 @@ class SACTest extends BaseSpecification {
         "Verify correct counts are returned by GetSummaryCounts"
         assert result.getNumDeployments() == 1
         assert result.getNumSecrets() == orchestrator.getSecretCount(DEPLOYMENT_QA1.namespace)
-        if (summaryTestShouldSeeNoClustersAndNodes()) {
-            assert result.getNumNodes() == 0
-            assert result.getNumClusters() == 0
-        }
         assert result.getNumImages() == 1
         cleanup:
         "Cleanup"
@@ -586,6 +660,17 @@ class SACTest extends BaseSpecification {
                 allAccessFlows.size() - allAccessFlowsWithoutNeighbors.size()
         assert sacFlowsNoQuery.size() - sacFlowsNoQueryFiltered.size() ==
                 allAccessFlows.size() - allAccessFlowsWithoutNeighbors.size()
+    }
+
+    @IgnoreIf({ Env.CI_JOBNAME.contains("postgres") })
+    def "test role aggregation should not combine permissions sets"() {
+        when:
+        useToken("aggregatedToken")
+
+        then:
+        def result = DeploymentService.listDeployments()
+        assert result.find { it.name == DEPLOYMENT_QA2.name }
+        assert !result.find { it.name == DEPLOYMENT_QA1.name }
     }
 
     private static List<DeploymentOuterClass.ListDeployment> listDeployments() {
